+++ ir/adt/pqueue.h
@@ -30,7 +30,7 @@ void del_pqueue(pqueue *q);
  * @param data   The actual data which should be stored in the queue.
  * @param key    The priority for the data.
  */
-void pqueue_put(pqueue *q, void *data, double key);
+void pqueue_put(pqueue *q, void *data, int key);
 
 /**
  * Returns and removes the first element, ie. that one with the highest priority, from the queue.
